Configure-ize the build
Similar to autoconf, generate output files from .in files and replace
any parameters specified on the command line in the .in files.

This removes the need to have the different plugconfX.rpgle files, though
not all options in plugconf.rpgle are available from configure - for those
that are not, you will need to patch (or manually edit) plugconf.rpgle.